Redesign DGKernels
