+++ framework/include/dgkernels/ADDGDiffusion.h
@@ -35,10 +35,7 @@ public:
   ADDGDiffusion(const InputParameters & parameters);
 
 protected:
-  virtual ADResidual computeADQpResidual(Moose::DGResidualType type) override;
-  virtual Real computeQpResidual(Moose::DGResidualType /*type*/) override { return 0.0; }
-  virtual Real computeQpJacobian(Moose::DGJacobianType /*type*/) override { return 0.0; }
-  // virtual Real computeQpJacobian(Moose::DGJacobianType type) override;
+  virtual ADResidual computeQpResidual(Moose::DGResidualType type) override;
 
   Real _epsilon;
   Real _sigma;
