+++ framework/src/dgkernels/DGKernel.C
@@ -27,157 +27,12 @@ template <>
 InputParameters
 validParams<DGKernel>()
 {
-  InputParameters params = validParams<MooseObject>();
-  params += validParams<TwoMaterialPropertyInterface>();
-  params += validParams<TransientInterface>();
-  params += validParams<BlockRestrictable>();
-  params += validParams<BoundaryRestrictable>();
-  params += validParams<MeshChangedInterface>();
-  params += validParams<TaggingInterface>();
-  params.addRequiredParam<NonlinearVariableName>(
-      "variable", "The name of the variable that this boundary condition applies to");
-  params.addParam<bool>("use_displaced_mesh",
-                        false,
-                        "Whether or not this object should use the "
-                        "displaced mesh for computation. Note that in "
-                        "the case this is true but no displacements "
-                        "are provided in the Mesh block the "
-                        "undisplaced mesh will still be used.");
-  params.addParamNamesToGroup("use_displaced_mesh", "Advanced");
-
-  params.declareControllable("enable");
+  InputParameters params = validParams<DGKernelBase>();
   params.registerBase("DGKernel");
-  params.addParam<std::vector<AuxVariableName>>(
-      "save_in",
-      "The name of auxiliary variables to save this Kernel's residual contributions to. "
-      " Everything about that variable must match everything about this variable (the "
-      "type, what blocks it's on, etc.)");
-  params.addParam<std::vector<AuxVariableName>>(
-      "diag_save_in",
-      "The name of auxiliary variables to save this Kernel's diagonal Jacobian "
-      "contributions to. Everything about that variable must match everything "
-      "about this variable (the type, what blocks it's on, etc.)");
-
   return params;
 }
 
-// Static mutex definitions
-Threads::spin_mutex DGKernel::_resid_vars_mutex;
-Threads::spin_mutex DGKernel::_jacoby_vars_mutex;
-
-DGKernel::DGKernel(const InputParameters & parameters)
-  : MooseObject(parameters),
-    BlockRestrictable(this),
-    BoundaryRestrictable(this, false), // false for _not_ nodal
-    SetupInterface(this),
-    TransientInterface(this),
-    FunctionInterface(this),
-    UserObjectInterface(this),
-    NeighborCoupleableMooseVariableDependencyIntermediateInterface(this, false, false),
-    NeighborMooseVariableInterface(
-        this, false, Moose::VarKindType::VAR_NONLINEAR, Moose::VarFieldType::VAR_FIELD_STANDARD),
-    TwoMaterialPropertyInterface(this, blockIDs(), boundaryIDs()),
-    Restartable(this, "DGKernels"),
-    MeshChangedInterface(parameters),
-    TaggingInterface(this),
-    _subproblem(*getCheckedPointerParam<SubProblem *>("_subproblem")),
-    _sys(*getCheckedPointerParam<SystemBase *>("_sys")),
-    _tid(parameters.get<THREAD_ID>("_tid")),
-    _assembly(_subproblem.assembly(_tid)),
-    _var(*mooseVariable()),
-    _mesh(_subproblem.mesh()),
-
-    _current_elem(_assembly.elem()),
-    _current_elem_volume(_assembly.elemVolume()),
-
-    _neighbor_elem(_assembly.neighbor()),
-
-    _current_side(_assembly.side()),
-    _current_side_elem(_assembly.sideElem()),
-    _current_side_volume(_assembly.sideElemVolume()),
-
-    _coord_sys(_assembly.coordSystem()),
-    _q_point(_assembly.qPointsFace()),
-    _qrule(_assembly.qRuleFace()),
-    _JxW(_assembly.JxWFace()),
-    _coord(_assembly.coordTransformation()),
-
-    _u(_is_implicit ? _var.sln() : _var.slnOld()),
-    _grad_u(_is_implicit ? _var.gradSln() : _var.gradSlnOld()),
-
-    _phi(_assembly.phiFace(_var)),
-    _grad_phi(_assembly.gradPhiFace(_var)),
-
-    _test(_var.phiFace()),
-    _grad_test(_var.gradPhiFace()),
-
-    _normals(_var.normals()),
-
-    _phi_neighbor(_assembly.phiFaceNeighbor(_var)),
-    _grad_phi_neighbor(_assembly.gradPhiFaceNeighbor(_var)),
-
-    _test_neighbor(_var.phiFaceNeighbor()),
-    _grad_test_neighbor(_var.gradPhiFaceNeighbor()),
-
-    _u_neighbor(_is_implicit ? _var.slnNeighbor() : _var.slnOldNeighbor()),
-    _grad_u_neighbor(_is_implicit ? _var.gradSlnNeighbor() : _var.gradSlnOldNeighbor()),
-
-    _save_in_strings(parameters.get<std::vector<AuxVariableName>>("save_in")),
-    _diag_save_in_strings(parameters.get<std::vector<AuxVariableName>>("diag_save_in"))
-{
-  addMooseVariableDependency(mooseVariable());
-
-  _save_in.resize(_save_in_strings.size());
-  _diag_save_in.resize(_diag_save_in_strings.size());
-
-  for (unsigned int i = 0; i < _save_in_strings.size(); i++)
-  {
-    MooseVariableFEBase * var = &_subproblem.getVariable(_tid,
-                                                         _save_in_strings[i],
-                                                         Moose::VarKindType::VAR_AUXILIARY,
-                                                         Moose::VarFieldType::VAR_FIELD_STANDARD);
-
-    if (_sys.hasVariable(_save_in_strings[i]))
-      mooseError("Trying to use solution variable " + _save_in_strings[i] +
-                 " as a save_in variable in " + name());
-
-    if (var->feType() != _var.feType())
-      paramError(
-          "save_in",
-          "saved-in auxiliary variable is incompatible with the object's nonlinear variable: ",
-          moose::internal::incompatVarMsg(*var, _var));
-
-    _save_in[i] = var;
-    var->sys().addVariableToZeroOnResidual(_save_in_strings[i]);
-    addMooseVariableDependency(var);
-  }
-
-  _has_save_in = _save_in.size() > 0;
-
-  for (unsigned int i = 0; i < _diag_save_in_strings.size(); i++)
-  {
-    MooseVariableFEBase * var = &_subproblem.getVariable(_tid,
-                                                         _diag_save_in_strings[i],
-                                                         Moose::VarKindType::VAR_NONLINEAR,
-                                                         Moose::VarFieldType::VAR_FIELD_STANDARD);
-
-    if (_sys.hasVariable(_diag_save_in_strings[i]))
-      mooseError("Trying to use solution variable " + _diag_save_in_strings[i] +
-                 " as a diag_save_in variable in " + name());
-
-    if (var->feType() != _var.feType())
-      paramError(
-          "diag_save_in",
-          "saved-in auxiliary variable is incompatible with the object's nonlinear variable: ",
-          moose::internal::incompatVarMsg(*var, _var));
-
-    _diag_save_in[i] = var;
-    var->sys().addVariableToZeroOnJacobian(_diag_save_in_strings[i]);
-    addMooseVariableDependency(var);
-  }
-
-  _has_diag_save_in = _diag_save_in.size() > 0;
-}
+DGKernel::DGKernel(const InputParameters & parameters) : DGKernelBase(parameters) {}
 
 DGKernel::~DGKernel() {}
 
@@ -213,16 +68,6 @@ DGKernel::computeElemNeighResidual(Moose::DGResidualType type)
       var->sys().solution().add_vector(_local_re, dof_indices);
     }
   }
-}
-
-void
-DGKernel::computeResidual()
-{
-  // Compute the residual for this element
-  computeElemNeighResidual(Moose::Element);
-
-  // Compute the residual for the neighbor
-  computeElemNeighResidual(Moose::Neighbor);
 }
 
 void
@@ -262,21 +107,6 @@ DGKernel::computeElemNeighJacobian(Moose::DGJacobianType type)
     }
   }
 }
-void
-DGKernel::computeJacobian()
-{
-  // Compute element-element Jacobian
-  computeElemNeighJacobian(Moose::ElementElement);
-
-  // Compute element-neighbor Jacobian
-  computeElemNeighJacobian(Moose::ElementNeighbor);
-
-  // Compute neighbor-element Jacobian
-  computeElemNeighJacobian(Moose::NeighborElement);
-
-  // Compute neighbor-neighbor Jacobian
-  computeElemNeighJacobian(Moose::NeighborNeighbor);
-}
 
 void
 DGKernel::computeOffDiagElemNeighJacobian(Moose::DGJacobianType type, unsigned int jvar)
@@ -299,47 +129,8 @@ DGKernel::computeOffDiagElemNeighJacobian(Moose::DGJacobianType type, unsigned i
   accumulateTaggedLocalMatrix();
 }
 
-void
-DGKernel::computeOffDiagJacobian(unsigned int jvar)
-{
-  if (jvar == _var.number())
-    computeJacobian();
-  else
-  {
-    // Compute element-element Jacobian
-    computeOffDiagElemNeighJacobian(Moose::ElementElement, jvar);
-
-    // Compute element-neighbor Jacobian
-    computeOffDiagElemNeighJacobian(Moose::ElementNeighbor, jvar);
-
-    // Compute neighbor-element Jacobian
-    computeOffDiagElemNeighJacobian(Moose::NeighborElement, jvar);
-
-    // Compute neighbor-neighbor Jacobian
-    computeOffDiagElemNeighJacobian(Moose::NeighborNeighbor, jvar);
-  }
-}
-
 Real
 DGKernel::computeQpOffDiagJacobian(Moose::DGJacobianType /*type*/, unsigned int /*jvar*/)
 {
   return 0.;
-}
-
-MooseVariable &
-DGKernel::variable()
-{
-  return _var;
-}
-
-SubProblem &
-DGKernel::subProblem()
-{
-  return _subproblem;
-}
-
-const Real &
-DGKernel::getNeighborElemVolume()
-{
-  return _assembly.neighborVolume();
 }
