+++ framework/src/dgkernels/DGKernelBase.C
@@ -1,0 +1,246 @@
+//* This file is part of the MOOSE framework
+//* https://www.mooseframework.org
+//*
+//* All rights reserved, see COPYRIGHT for full restrictions
+//* https://github.com/idaholab/moose/blob/master/COPYRIGHT
+//*
+//* Licensed under LGPL 2.1, please see LICENSE for details
+//* https://www.gnu.org/licenses/lgpl-2.1.html
+
+#include "DGKernelBase.h"
+#include "Assembly.h"
+#include "MooseVariable.h"
+#include "Problem.h"
+#include "SubProblem.h"
+#include "SystemBase.h"
+#include "MaterialData.h"
+#include "ParallelUniqueId.h"
+
+#include "libmesh/dof_map.h"
+#include "libmesh/dense_vector.h"
+#include "libmesh/numeric_vector.h"
+#include "libmesh/dense_subvector.h"
+#include "libmesh/libmesh_common.h"
+#include "libmesh/quadrature.h"
+
+template <>
+InputParameters
+validParams<DGKernelBase>()
+{
+  InputParameters params = validParams<MooseObject>();
+  params += validParams<TwoMaterialPropertyInterface>();
+  params += validParams<TransientInterface>();
+  params += validParams<BlockRestrictable>();
+  params += validParams<BoundaryRestrictable>();
+  params += validParams<MeshChangedInterface>();
+  params += validParams<TaggingInterface>();
+  params.addRequiredParam<NonlinearVariableName>(
+      "variable", "The name of the variable that this boundary condition applies to");
+  params.addParam<bool>("use_displaced_mesh",
+                        false,
+                        "Whether or not this object should use the "
+                        "displaced mesh for computation. Note that in "
+                        "the case this is true but no displacements "
+                        "are provided in the Mesh block the "
+                        "undisplaced mesh will still be used.");
+  params.addParamNamesToGroup("use_displaced_mesh", "Advanced");
+
+  params.declareControllable("enable");
+  params.addParam<std::vector<AuxVariableName>>(
+      "save_in",
+      "The name of auxiliary variables to save this Kernel's residual contributions to. "
+      " Everything about that variable must match everything about this variable (the "
+      "type, what blocks it's on, etc.)");
+  params.addParam<std::vector<AuxVariableName>>(
+      "diag_save_in",
+      "The name of auxiliary variables to save this Kernel's diagonal Jacobian "
+      "contributions to. Everything about that variable must match everything "
+      "about this variable (the type, what blocks it's on, etc.)");
+
+  return params;
+}
+
+// Static mutex definitions
+Threads::spin_mutex DGKernelBase::_resid_vars_mutex;
+Threads::spin_mutex DGKernelBase::_jacoby_vars_mutex;
+
+DGKernelBase::DGKernelBase(const InputParameters & parameters)
+  : MooseObject(parameters),
+    BlockRestrictable(this),
+    BoundaryRestrictable(this, false), // false for _not_ nodal
+    SetupInterface(this),
+    TransientInterface(this),
+    FunctionInterface(this),
+    UserObjectInterface(this),
+    NeighborCoupleableMooseVariableDependencyIntermediateInterface(this, false, false),
+    NeighborMooseVariableInterface(
+        this, false, Moose::VarKindType::VAR_NONLINEAR, Moose::VarFieldType::VAR_FIELD_STANDARD),
+    TwoMaterialPropertyInterface(this, blockIDs(), boundaryIDs()),
+    Restartable(this, "DGKernels"),
+    MeshChangedInterface(parameters),
+    TaggingInterface(this),
+    _subproblem(*getCheckedPointerParam<SubProblem *>("_subproblem")),
+    _sys(*getCheckedPointerParam<SystemBase *>("_sys")),
+    _tid(parameters.get<THREAD_ID>("_tid")),
+    _assembly(_subproblem.assembly(_tid)),
+    _var(*mooseVariable()),
+    _mesh(_subproblem.mesh()),
+
+    _current_elem(_assembly.elem()),
+    _current_elem_volume(_assembly.elemVolume()),
+
+    _neighbor_elem(_assembly.neighbor()),
+
+    _current_side(_assembly.side()),
+    _current_side_elem(_assembly.sideElem()),
+    _current_side_volume(_assembly.sideElemVolume()),
+
+    _coord_sys(_assembly.coordSystem()),
+    _q_point(_assembly.qPointsFace()),
+    _qrule(_assembly.qRuleFace()),
+    _JxW(_assembly.JxWFace()),
+    _coord(_assembly.coordTransformation()),
+
+    _u(_is_implicit ? _var.sln() : _var.slnOld()),
+    _grad_u(_is_implicit ? _var.gradSln() : _var.gradSlnOld()),
+
+    _phi(_assembly.phiFace(_var)),
+    _grad_phi(_assembly.gradPhiFace(_var)),
+
+    _test(_var.phiFace()),
+    _grad_test(_var.gradPhiFace()),
+
+    _normals(_var.normals()),
+
+    _phi_neighbor(_assembly.phiFaceNeighbor(_var)),
+    _grad_phi_neighbor(_assembly.gradPhiFaceNeighbor(_var)),
+
+    _test_neighbor(_var.phiFaceNeighbor()),
+    _grad_test_neighbor(_var.gradPhiFaceNeighbor()),
+
+    _u_neighbor(_is_implicit ? _var.slnNeighbor() : _var.slnOldNeighbor()),
+    _grad_u_neighbor(_is_implicit ? _var.gradSlnNeighbor() : _var.gradSlnOldNeighbor()),
+
+    _save_in_strings(parameters.get<std::vector<AuxVariableName>>("save_in")),
+    _diag_save_in_strings(parameters.get<std::vector<AuxVariableName>>("diag_save_in"))
+{
+  addMooseVariableDependency(mooseVariable());
+
+  _save_in.resize(_save_in_strings.size());
+  _diag_save_in.resize(_diag_save_in_strings.size());
+
+  for (unsigned int i = 0; i < _save_in_strings.size(); i++)
+  {
+    MooseVariableFEBase * var = &_subproblem.getVariable(_tid,
+                                                         _save_in_strings[i],
+                                                         Moose::VarKindType::VAR_AUXILIARY,
+                                                         Moose::VarFieldType::VAR_FIELD_STANDARD);
+
+    if (_sys.hasVariable(_save_in_strings[i]))
+      mooseError("Trying to use solution variable " + _save_in_strings[i] +
+                 " as a save_in variable in " + name());
+
+    if (var->feType() != _var.feType())
+      paramError(
+          "save_in",
+          "saved-in auxiliary variable is incompatible with the object's nonlinear variable: ",
+          moose::internal::incompatVarMsg(*var, _var));
+
+    _save_in[i] = var;
+    var->sys().addVariableToZeroOnResidual(_save_in_strings[i]);
+    addMooseVariableDependency(var);
+  }
+
+  _has_save_in = _save_in.size() > 0;
+
+  for (unsigned int i = 0; i < _diag_save_in_strings.size(); i++)
+  {
+    MooseVariableFEBase * var = &_subproblem.getVariable(_tid,
+                                                         _diag_save_in_strings[i],
+                                                         Moose::VarKindType::VAR_NONLINEAR,
+                                                         Moose::VarFieldType::VAR_FIELD_STANDARD);
+
+    if (_sys.hasVariable(_diag_save_in_strings[i]))
+      mooseError("Trying to use solution variable " + _diag_save_in_strings[i] +
+                 " as a diag_save_in variable in " + name());
+
+    if (var->feType() != _var.feType())
+      paramError(
+          "diag_save_in",
+          "saved-in auxiliary variable is incompatible with the object's nonlinear variable: ",
+          moose::internal::incompatVarMsg(*var, _var));
+
+    _diag_save_in[i] = var;
+    var->sys().addVariableToZeroOnJacobian(_diag_save_in_strings[i]);
+    addMooseVariableDependency(var);
+  }
+
+  _has_diag_save_in = _diag_save_in.size() > 0;
+}
+
+DGKernelBase::~DGKernelBase() {}
+
+void
+DGKernelBase::computeResidual()
+{
+  // Compute the residual for this element
+  computeElemNeighResidual(Moose::Element);
+
+  // Compute the residual for the neighbor
+  computeElemNeighResidual(Moose::Neighbor);
+}
+
+void
+DGKernelBase::computeJacobian()
+{
+  // Compute element-element Jacobian
+  computeElemNeighJacobian(Moose::ElementElement);
+
+  // Compute element-neighbor Jacobian
+  computeElemNeighJacobian(Moose::ElementNeighbor);
+
+  // Compute neighbor-element Jacobian
+  computeElemNeighJacobian(Moose::NeighborElement);
+
+  // Compute neighbor-neighbor Jacobian
+  computeElemNeighJacobian(Moose::NeighborNeighbor);
+}
+
+void
+DGKernelBase::computeOffDiagJacobian(unsigned int jvar)
+{
+  if (jvar == _var.number())
+    computeJacobian();
+  else
+  {
+    // Compute element-element Jacobian
+    computeOffDiagElemNeighJacobian(Moose::ElementElement, jvar);
+
+    // Compute element-neighbor Jacobian
+    computeOffDiagElemNeighJacobian(Moose::ElementNeighbor, jvar);
+
+    // Compute neighbor-element Jacobian
+    computeOffDiagElemNeighJacobian(Moose::NeighborElement, jvar);
+
+    // Compute neighbor-neighbor Jacobian
+    computeOffDiagElemNeighJacobian(Moose::NeighborNeighbor, jvar);
+  }
+}
+
+MooseVariable &
+DGKernelBase::variable()
+{
+  return _var;
+}
+
+SubProblem &
+DGKernelBase::subProblem()
+{
+  return _subproblem;
+}
+
+const Real &
+DGKernelBase::getNeighborElemVolume()
+{
+  return _assembly.neighborVolume();
+}
